Some renamings
