+++ frontend/lib/src/components/widgets/DataFrame/EditingState.test.ts
@@ -21,7 +21,7 @@ import {
   BaseColumnProps,
   TextColumn,
 } from "@streamlit/lib/src/components/widgets/DataFrame/columns"
-import { DataFrameCellType } from "@streamlit/lib/src/dataframes/arrowParseUtils"
+import { DataFrameCellType } from "@streamlit/lib/src/dataframes/arrowTypeUtils"
 
 import EditingState from "./EditingState"
