+++ frontend/lib/src/components/widgets/DataFrame/columns/ChartColumn.test.ts
@@ -18,7 +18,7 @@ import { GridCellKind } from "@glideapps/glide-data-grid"
 import { SparklineCellType } from "@glideapps/glide-data-grid-cells"
 import { Field, Float64, List } from "apache-arrow"
 
-import { DataFrameCellType } from "@streamlit/lib/src/dataframes/arrowParseUtils"
+import { DataFrameCellType } from "@streamlit/lib/src/dataframes/arrowTypeUtils"
 
 import {
   AREA_CHART_TYPE,
