+++ frontend/lib/src/components/widgets/DataFrame/columns/ImageColumn.test.ts
@@ -17,7 +17,7 @@
 import { GridCellKind, ImageCell } from "@glideapps/glide-data-grid"
 import { Field, Utf8 } from "apache-arrow"
 
-import { DataFrameCellType } from "@streamlit/lib/src/dataframes/arrowParseUtils"
+import { DataFrameCellType } from "@streamlit/lib/src/dataframes/arrowTypeUtils"
 
 import ImageColumn from "./ImageColumn"
