+++ frontend/lib/src/components/widgets/DataFrame/columns/ObjectColumn.test.ts
@@ -17,7 +17,7 @@
 import { GridCellKind, TextCell } from "@glideapps/glide-data-grid"
 import { Field, Utf8 } from "apache-arrow"
 
-import { DataFrameCellType } from "@streamlit/lib/src/dataframes/arrowParseUtils"
+import { DataFrameCellType } from "@streamlit/lib/src/dataframes/arrowTypeUtils"
 
 import ObjectColumn from "./ObjectColumn"
