+++ frontend/lib/src/components/widgets/DataFrame/columns/utils.ts
@@ -31,7 +31,7 @@ import "moment-timezone"
 import numbro from "numbro"
 import { sprintf } from "sprintf-js"
 
-import { ArrowType } from "@streamlit/lib/src/dataframes/arrowParseUtils"
+import { ArrowType } from "@streamlit/lib/src/dataframes/arrowTypeUtils"
 import { EmotionTheme } from "@streamlit/lib/src/theme"
 import {
   isNullOrUndefined,
