+++ frontend/lib/src/components/widgets/DataFrame/hooks/useColumnReordering.test.ts
@@ -22,7 +22,7 @@ import {
   NumberColumn,
   TextColumn,
 } from "@streamlit/lib/src/components/widgets/DataFrame/columns"
-import { DataFrameCellType } from "@streamlit/lib/src/dataframes/arrowParseUtils"
+import { DataFrameCellType } from "@streamlit/lib/src/dataframes/arrowTypeUtils"
 
 import useColumnReordering from "./useColumnReordering"
